requirements.txt
PyQt5==5.15.10
pyserial~=3.5